AMLS-4079 Moved migrator into sub-package
